backend/requirements.txt
# Django 5.x
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=65.0,<66.0
dj-rest-auth>=7.0,<8.0

# Redis/Celery
redis>=5.0,<6.0
celery>=5.3,<6.0

# Utils
python-decouple>=3.8,<4.0
Pillow>=10.0,<11.0
requests>=2.31,<3.0
openpyxl>=3.1,<4.0

# LGPD / Compliance
django-simple-history>=3.5,<4.0

# Dev & Debug
django-debug-toolbar>=4.3,<5.0
django-extensions>=3.2,<4.0

# Production
gunicorn>=21.0,<24.0
whitenoise>=6.6,<7.0
